remove deprecated apis
