+++ proto/server.proto
@@ -24,11 +24,6 @@ service Server {
   rpc GetStoreApps (GetStoreAppsRequest) returns (GetStoreAppsResponse) {}
   rpc GetStoreAppValues(GetStoreAppValuesRequest) returns (GetStoreAppValuesResponse) {}
 
-  rpc DeployStoreApp(DeployStoreAppRequest) returns (DeployStoreAppResponse) {}
-  rpc UnDeployStoreApp(UnDeployStoreAppRequest) returns (UnDeployStoreAppResponse) {}
-
-  rpc UpgradeStoreApp(UpgradeStoreAppRequest) returns (UpgradeStoreAppResponse) {}
-
   rpc StoreCredential (StoreCredentialRequest) returns (StoreCredentialResponse) {}
 }
 
@@ -254,30 +249,6 @@ message GetStoreAppValuesResponse {
   bytes overrideValues = 3;
 }
 
-message UpgradeStoreAppRequest {
-  string appName = 1;
-  string version = 2;
-  string clusterID = 3;
-  bytes overrideValues = 4;
-}
-
-message UpgradeStoreAppResponse {
-  StatusCode status = 1;
-  string statusMessage = 2;
-}
-
-message DeployStoreAppRequest {
-  string appName = 1;
-  string version = 2;
-  string clusterID = 3;
-  bytes overrideValues = 4;
-}
-
-message DeployStoreAppResponse{
-  StatusCode status = 1;
-  string statusMessage = 2;
-}
-
 message StoreAppConfig {
   string releaseName = 1;
   string appName = 2;
@@ -317,12 +288,3 @@ message StoreCredentialResponse {
   string statusMessage = 2;
 }
 
-message UnDeployStoreAppRequest {
-  string clusterID = 1;
-  string releaseName = 5;
-}
-
-message UnDeployStoreAppResponse{
-  StatusCode status = 1;
-  string statusMessage = 2;
-}
